create a webpack build process for node modules, add libs to package.json, reorganize around the src folder
