+++ inc/enqueue.php
@@ -25,19 +25,19 @@ if ( ! function_exists( 'understrap_scripts' ) ) {
 
 		$js_version = $theme_version . '.' . filemtime( get_template_directory() . '/js/theme.min.js' );
 		wp_enqueue_script( 'understrap-scripts', get_template_directory_uri() . '/js/theme.min.js', array(), $js_version, true );
-		
+
 		if ( is_singular() && comments_open() && get_option( 'thread_comments' ) ) {
 			wp_enqueue_script( 'comment-reply' );
 		}
-		if($post->post_type == 'portfolio'){		
-			wp_enqueue_script( 'envs-assess', get_template_directory_uri() . '/js/envs-assess.js', array(), '1', true );
+		if($post->post_type == 'portfolio'){
+			wp_enqueue_script( 'envs-assess', get_template_directory_uri() . '/js/app.js', array(), '1', true );
 			$port_url = get_field('portfolio_url', $post->ID);
 			wp_localize_script('envs-assess', 'port', array(
 			    'url' => $port_url,
 			));
 		}
-		if(get_page_template_slug($post->ID)== 'page-templates/right-sidebarpage-assessment.php'){		
-			wp_enqueue_script( 'envs-assess-item', get_template_directory_uri() . '/js/envs-assess-item.js', array(), '1', true );
+		if(get_page_template_slug($post->ID)== 'page-templates/right-sidebarpage-assessment.php'){
+			wp_enqueue_script( 'envs-assess-item', get_template_directory_uri() . '/js/app.js', array(), '1', true );
 			// $port_url = get_field('portfolio_url', $post->ID);
 			// wp_localize_script('envs-assess', 'port', array(
 			//     'url' => $port_url,
