+++ js/app.js
@@ -1,0 +1,1 @@
+!function(e){var t={};function n(r){if(t[r])return t[r].exports;var o=t[r]={i:r,l:!1,exports:{}};return e[r].call(o.exports,o,o.exports,n),o.l=!0,o.exports}n.m=e,n.c=t,n.d=function(e,t,r){n.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:r})},n.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},n.t=function(e,t){if(1&t&&(e=n(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(n.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var o in e)n.d(r,o,function(t){return e[t]}.bind(null,o));return r},n.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return n.d(t,"a",t),t},n.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},n.p="",n(n.s=2)}([function(e,t,n){"use strict";Object.defineProperty(t,"__esModule",{value:!0});var r=function(e){return(e.split(/\s+/)||[""]).length},o=function(e){return function(e){var t=e.toLowerCase().replace(/'/g,"").replace(/e\b/g,""),n=t.split(/[^aeiouy]+/).filter(Boolean);return null==e||""===e?0:1===t.length?1:[/cial/,/tia/,/cius/,/cious/,/giu/,/ion/,/iou/,/sia$/,/.ely$/,/sed$/].map((function(e){return(t.match(e)||[]).length})).reduce((function(e,t){return e-t}))+[/ia/,/riet/,/dien/,/iu/,/io/,/ii/,/[aeiouym]bl$/,/[aeiou]{3}/,/^mc/,/ism$/,/([^aeiouy])\1l$/,/[^l]lien/,/^coa[dglx]./,/[^gq]ua[^auieo]/,/dnt$/].map((function(e){return(t.match(e)||[]).length})).reduce((function(e,t){return e+t}))+n.length-(n.length>0&&""===n[0]?1:0)+t.split(/\b/).map((function(e){return e.trim()})).filter(Boolean).filter((function(e){return!e.match(/[.,'!?]/g)})).map((function(e){return e.match(/[aeiouy]/)?0:1})).reduce((function(e,t){return e+t}))}(e)/r(e)},u=function(e){return r(e)/function(e){return(e.split(". ")||[""]).length}(e)};t.rate=function(e){return 206.835-1.015*u(e)-84.6*o(e)},t.grade=function(e){return.39*u(e)+11.8*o(e)-15.59}},function(e,t,n){!function(t){let n=[];const r=Array.prototype.forEach;function o(e){const t=[];let n=0;const r=e.length;for(;n<r;){const o=e.charCodeAt(n++);if(o>=55296&&o<=56319&&n<r){const r=e.charCodeAt(n++);56320==(64512&r)?t.push(((1023&o)<<10)+(1023&r)+65536):(t.push(o),n--)}else t.push(o)}return t}function u(e,t){const n=Object.prototype.toString.call(e),r="string"==typeof e||"[object NodeList]"===n||"[object HTMLCollection]"===n||1===e.nodeType,o="function"==typeof t;return r||console.error("Countable: Not a valid target"),o||console.error("Countable: Not a valid callback function"),r&&o}function i(e,t){let n=""+("string"==typeof e?e:"value"in e?e.value:e.textContent);(t=t||{}).stripTags&&(n=n.replace(/<\/?[a-z][^>]*>/gi,"")),t.ignore&&r.call(t.ignore,(function(e){n=n.replace(e,"")}));const u=n.trim();return{paragraphs:u?(u.match(t.hardReturns?/\n{2,}/g:/\n+/g)||[]).length+1:0,sentences:u?(u.match(/[.?!…]+./g)||[]).length+1:0,words:u?(u.replace(/['";:,.?¿\-!¡]+/g,"").match(/\S+/g)||[]).length:0,characters:u?o(u.replace(/\s/g,"")).length:0,all:o(n).length}}const s={on:function(e,t,o){if(u(e,t))return Array.isArray(e)||(e=[e]),r.call(e,(function(e){const r=function(){t.call(e,i(e,o))};n.push({element:e,handler:r}),r(),e.addEventListener("input",r)})),this},off:function(e){if(u(e,(function(){})))return Array.isArray(e)||(e=[e]),n.filter((function(t){return-1!==e.indexOf(t.element)})).forEach((function(e){e.element.removeEventListener("input",e.handler)})),n=n.filter((function(t){return-1===e.indexOf(t.element)})),this},count:function(e,t,n){if(u(e,t))return Array.isArray(e)||(e=[e]),r.call(e,(function(e){t.call(e,i(e,n))})),this},enabled:function(e){return void 0===e.length&&(e=[e]),n.filter((function(t){return-1!==e.indexOf(t.element)})).length===e.length}};e.exports=s}()},function(e,t,n){"use strict";n.r(t);var r=n(1),o=n.n(r),u=n(0);console.log("envs-assess"),console.log(port.url);let i=port.url+"/wp-json";function s(e,t){console.log(e.length),e.length>1&&(t+="s");let n=e.join(", ");return`<div class="${t.toLowerCase()}"><h3>${t}</h3><p>${n}</p></div>`}console.log(i),function(){let e=port.url+"/wp-json";jQuery.get(e,(function(e){jQuery("#bio-data");let t=e.envs.student_name,n=e.url,r=e.envs.bio_pic.thumbnail,o=e.envs.graduation_date,u=s(e.envs.majors,"Major"),i=s(e.envs.minors,"Minor"),c=e.post_count,a=e.page_count.publish,l=e.created.substring(0,10),d=e.last_updated.substring(0,10),f=document.getElementById("bioModalLabel"),p=document.getElementById("stu-photo"),g=document.getElementById("stu-grad"),h=document.getElementById("stu-majors"),m=document.getElementById("stu-minors"),y=document.getElementById("stu-post-count"),b=document.getElementById("stu-page-count"),v=document.getElementById("stu-created"),j=document.getElementById("stu-last-update");f.innerHTML=`<a href="${n}">${t}</a>`,p.src=r,g.innerHTML=`<h3>Graduation</h3> <p>${o}</p>`,h.innerHTML=u,m.innerHTML=i,y.innerHTML=`<p>${c} Posts</p>`,b.innerHTML=`<p>${a} Pages</p>`,v.innerHTML=`<p>Created on ${l}</p>`,j.innerHTML=`<p>Updated on ${d}</p>`}),"json")}(),console.log("foo");var c="The quick brown fox jumped over the lazy dogs";console.log(Object(u.rate)(c));c="The quick brown fox jumped over the lazy dogs";console.log(Object(u.grade)(c)),function(){let e=port.url+"/wp-json/wp/v2/posts";jQuery(".spinner-image").show(),jQuery.get(e,(function(e){e.forEach((function(e){jQuery("#envs-post-data").append(function(e){let t=e.title.rendered,n=e.guid.rendered;return`<h2><a href="${n}">${t}</a></h2><a class="rubric-button" href="http://192.168.33.10/wordpress/envs-assess/rubric-one/?assess-source=${n}">Rubric 1</a><a class="rubric-button" href="http://192.168.33.10/wordpress/envs-assess/rubric-two/?assess-source=${n}">Rubric 2</a>`}(e));const t=e.content.rendered;o.a.count(t,e=>(function(e){let t=e.paragraphs,n=e.sentences,r=e.words;return jQuery("#envs-post-data").append(`<div class="envs-post-stats">${t} paragraphs<br> ${n} sentences<br> ${r} words</div>`)})(e))}))}),"json")}()}]);
