+++ package.json
@@ -30,6 +30,7 @@
     "@babel/preset-env": "^7.6.2",
     "bootstrap": "^4.3.1",
     "browser-sync": "^2.26.7",
+    "countable": "^3.0.1",
     "del": "^4.1.1",
     "flesch-kincaid": "github:daveross/flesch-kincaid",
     "font-awesome": "^4.7.0",
@@ -47,6 +48,7 @@
     "gulp-sequence": "^1.0.0",
     "gulp-sourcemaps": "^2.6.5",
     "gulp-uglify": "^3.0.2",
+    "gulp-uglify-es": "^1.0.4",
     "gulp-watch": "^5.0.1",
     "run-sequence": "^2.2.1",
     "undescores-for-npm": "^1.0.0"
@@ -55,6 +57,8 @@
     "@babel/core": "^7.6.2",
     "autoprefixer": "^9.6.4",
     "gulp-babel": "^8.0.0",
-    "gulp-postcss": "^8.0.0"
+    "gulp-postcss": "^8.0.0",
+    "webpack": "^4.41.0",
+    "webpack-stream": "^5.2.1"
   }
 }
